+++ test/integration/8453/PricingAdapter/GetLeverageTokenPriceInCollateral.t.sol
@@ -24,7 +24,7 @@ contract GetLeverageTokenPriceInCollateralTest is PricingAdapterTest {
         uint256 sharesToMint = 1e18;
         _mint(user, sharesToMint, type(uint256).max);
 
-        uint256 managementFee = 10_00; // 10%
+        uint16 managementFee = 10_00; // 10%
         leverageManager.setManagementFee(leverageToken, managementFee);
 
         uint256 resultA = pricingAdapter.getLeverageTokenPriceInCollateral(leverageToken);
