+++ test/unit/LeverageManager/ConvertSharesToCollateral.t.sol
@@ -212,7 +212,7 @@ contract ConvertSharesToCollateralTest is LeverageManagerTest {
         uint128 totalSupply = 100;
         uint128 totalCollateral = 99;
 
-        uint256 managementFee = 0.5e4; // 50%
+        uint16 managementFee = 0.5e4; // 50%
         _setManagementFee(feeManagerRole, leverageToken, managementFee);
 
         lendingAdapter.mockCollateral(totalCollateral);
